likes feature works

+++ src/app/topic-page/topic-list/topic-list.component.ts
@@ -1,6 +1,6 @@
-import { Component, OnInit, Output, EventEmitter } from '@angular/core';
+import { Component, OnInit, Output, EventEmitter, Input } from '@angular/core';
 import { TopicDTO } from '../model/TopicDTO';
-import { UserTopicDTO } from '../model/UserTopicDTO';
+import { TopicWithUserDetails } from '../model/UserTopicDTO';
 import { UserDTO } from '../model/UserDTO';
 import { TopicService } from '../services/topic-service/topic-service.service';
 import { UserService } from 'src/app/global-services/user-service/user-service.service';
@@ -12,20 +12,14 @@ import { LoginService } from 'src/app/global-services/login-service/login-servic
   styleUrls: ['./topic-list.component.css']
 })
 export class TopicListComponent implements OnInit {
-  userList: UserDTO[] = [];
-  topicList: TopicDTO[] = [];
-  topicUserList: UserTopicDTO[] = [];
+  @Input()
+  topicUserList: TopicWithUserDetails[];
 
   @Output()
-  userTopicEmitter: EventEmitter<any> = new EventEmitter();
+  loadTopicEmitter: EventEmitter<any> = new EventEmitter();
 
-  topicTitleText: string = '';
-  topicTitleFlag: boolean = false;
-  postContentText: string = '';
-  postContentFlag: boolean = false;
   constructor(
     private topicService: TopicService,
-    private userService: UserService,
     private loginService: LoginService
   ) {}
 
@@ -34,52 +28,26 @@ export class TopicListComponent implements OnInit {
   }
 
   loadTopics() {
-    this.topicList = [];
-    this.userList = [];
-    this.topicUserList = [];
-    this.topicService.getTopicList().subscribe((topicResponse: TopicDTO[]) => {
-      let userUuids = new Set<string>();
-      this.topicList = topicResponse;
-
-      this.topicList.forEach(topic => {
-        userUuids.add(topic.authorUuid);
-      });
-
-      this.userService
-        .getUsersListByUuidList(Array.from(userUuids.values()))
-        .subscribe((response: UserDTO[]) => {
-          this.userList = response;
-
-          this.topicList.forEach(topic => {
-            let user: UserDTO = null;
-
-            this.userList.forEach(u => {
-              if (u.uuid === topic.authorUuid) {
-                user = u;
-              }
-            });
-
-            let userWithTopic: UserTopicDTO = {
-              author: user,
-              topic: topic
-            };
-            this.topicUserList.push(userWithTopic);
-          });
-
-          this.userTopicEmitter.emit(this.topicUserList);
-        });
-    });
+    this.loadTopicEmitter.emit(true);
   }
 
   deleteTopic(topic: TopicDTO) {
-    this.topicService
-      .deletePostByUuid(topic.uuid)
-      .subscribe((response: TopicDTO) => {
-        this.loadTopics();
-      });
+    let confirmation = confirm(
+      'Are you sure to delete this topic "' +
+        this.cutTooLongTitles(topic.title) +
+        '" ?'
+    );
+
+    if (confirmation === true) {
+      this.topicService
+        .deletePostByUuid(topic.uuid)
+        .subscribe((response: TopicDTO) => {
+          this.loadTopicEmitter.emit(true);
+        });
+    }
   }
 
-  isTopicDeletable(authorUuid: string) {
+  isTopicRemovable(authorUuid: string) {
     return this.loginService.loggedUser.uuid === authorUuid;
   }
